Implement Decision Matrix mutation (#33)
* Implement Decision Matrix mutation (#12)

Wires the five DM mutation operators from #1 §5 (reroll action, reroll
mode, reroll sensor, nudge threshold, reroll jump offset) into
MutationServiceImpl alongside the existing trait mutation. SplitStep's
offspring already call mutatedDna, so they now get a mutated copy of the
parent's DM for free instead of an exact copy.

Extra guidelines from the issue thread:
- Mutation is now a two-stage roll: MutationRate gates whether a mutation
  happens at all (unchanged), then a second roll picks DM vs. trait,
  weighted by the new Settings.DmMutationRatio (default 0.8, i.e. 80% DM).
- Added a DM Mutation Ratio slider (SlidersStage) and DM/Trait mutation
  counters (StatisticsStage), backed by MutationService.dmMutationCount /
  traitMutationCount.

DecisionMatrix gains withInstruction(index, instruction) so a mutation
can rewrite exactly one state immutably. RandomService gains a
random(Float, Float) overload for the threshold nudge.

Test coverage for each operator, the DM-vs-trait routing, the counters,
and the wrap-on-mutation behavior for jump offsets (issue's own example:
an offset of 30 on a 25-cell matrix wraps to 5).


* Track mutation counts per cell instead of globally on the service

Moves dmMutationCount/traitMutationCount from MutationService (a global,
simulation-wide tally) onto DNA itself: each genome now carries its own
count of DM and trait mutations accumulated since the founder genome that
seeded the population - i.e. that cell's "distance" from the original.
MutationServiceImpl.mutatedDna() bumps the relevant count on the returned
DNA instead of an internal service field, so the count travels with the
genome through SplitStep's inheritance chain automatically.

Moves the display from StatisticsStage (a global counter didn't fit the
new per-lineage semantics) to CellDetailsStage, alongside the existing
per-cell Position/Energy/Size/Age fields for the selected cell.


* Change theshold to Int. Fix mutation logic

* Fix display details

* Fix tests for the Int-percentage threshold refactor, fix two bugs found in review

Fixes all tests broken by the Instruction.threshold Double->Int change
(now integer percentages/counts rather than 0.0-1.0 ratios), and by
randomJumpOffset() drawing directly from (0, SIZE-1) instead of a wider
range that needed wrapping.

Also fixes two bugs surfaced while reviewing the refactor:
- nudgedThreshold's small-value fallback (offsetPercents sign as a unit
  step) bypassed coerceAtLeast(0) on early-return paths, so a threshold
  already at 0 could nudge to -1. Restructured so the floor applies on
  every path.
- Organic.toPercentage divided by `whole` unconditionally; energyPercentage
  divides by `size`, which drain() can (rarely, and normally alongside
  clamping energy to match) take to 0. Guards against a division-by-zero
  crash regardless of whether that invariant holds everywhere.

Also updates the stale KDoc on directionAwayFromToxin/directionAwayFromCrowd
(still described a nullable/random-fallback contract that no longer
applies since the field-boundary check was added), and adds test coverage
for that new boundary check.


* Rewrite nudgedThreshold to sample from a window instead of applying a delta

Replaces the offset-then-clamp approach (which needed a special case for
small thresholds rounding to a zero offset) with directly sampling a new
value from [threshold - offsetRange, threshold + offsetRange], with the
low end floored at 0 and the window's half-width floored at
MINIMAL_THRESHOLD_NUDGE_RANGE. Simpler, and the floor is now structural
(part of the range itself) rather than a separate branch that could be
bypassed.


---------

+++ core/src/main/com/invenit/bacillus/model/matrix/Instruction.kt
@@ -9,6 +9,6 @@ data class Instruction(
     val action: Action,
     val sensor: Sensor,
     val comparator: Comparator,
-    val threshold: Double,
+    val threshold: Int,
     val jumpOffset: Int
 )
